Check removed resources in API update
